Update copyright, versioning, change directory structure
