+++ LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2021- Sanjay-B(Sanjay Bhadra)
+Copyright (c) 2024- sbhadr (Sanjay Bhadra)
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
