--- pyblox3/__init__.py
@@ -1,29 +1,0 @@
-#
-#  __init__.py
-#  pyblox
-#
-#  By Sanjay-B(Sanjay Bhadra)
-#  Copyright © 2019- Sanjay-B(Sanjay Bhadra). All rights reserved.
-#
-
-__title__ = 'pyblox'
-__author__ = 'Sanjay-B'
-__license__ = 'MIT'
-__copyright__ = 'Copyright © 2021- Sanjay-B(Sanjay Bhadra)'
-
-# Dependencies
-
-
-# Parent Class Modules
-from .src.groups import *
-from .src.users import *
-from .src.auth import *
-from .src.abtesting import *
-from .src.accountInformation import *
-from .src.privateMessages import *
-from .src.games import *
-
-
-
-# Utility Modules
-from .src.util import *
