--- pyblox3/src/util/__init__.py
@@ -1,19 +1,0 @@
-#
-#  __init__.py
-#  pyblox -> util
-#
-#  By Sanjay-B(Sanjay Bhadra)
-#  Copyright © 2019- Sanjay-B(Sanjay Bhadra). All rights reserved.
-#
-
-# Some nifty methods that I like to keep handy because
-# I write them over and over for every project.
-# Bet you relate.. 
-#
-# Message from Sanjay 2024: Who the hell uses
-# the word nifty..?
-
-# Parent Class Modules
-from .req import Req
-from .io import io
-from .cached import Cached
