--- pyblox3/src/util/cached.py
@@ -1,12 +1,0 @@
-#
-#  util -> cached.py
-#  pyblox
-#
-#  By Sanjay-B(Sanjay Bhadra)
-#  Copyright © 2019- Sanjay-B(Sanjay Bhadra). All rights reserved.
-#
-
-class Cached:
-	cache = {}
-	file_cache = {}
-	auth = {}
