+++ src/webapi/groups.py
@@ -2,11 +2,11 @@
 #  groups.py
 #  pyblox
 #
-#  By Sanjay-B(Sanjay Bhadra)
-#  Copyright © 2019- Sanjay-B(Sanjay Bhadra). All rights reserved.
+#  By sbhadr (Sanjay Bhadra)
+#  Copyright © 2024- sbhadr (Sanjay Bhadra). All rights reserved.
 #
 
-from .util import *
+from .util import Req
 
 Req = Req()
