+++ src/webapi/privateMessages.py
@@ -2,11 +2,11 @@
 #  privateMessages.py
 #  pyblox
 #
-#  By Sanjay-B(Sanjay Bhadra)
-#  Copyright © 2021- Sanjay-B(Sanjay Bhadra). All rights reserved.
+#  By sbhadr (Sanjay Bhadra)
+#  Copyright © 2024- sbhadr (Sanjay Bhadra). All rights reserved.
 #
 
-from .util import *
+from .util import Req
 
 Req = Req()
