+++ src/webapi/util/__init__.py
@@ -1,0 +1,19 @@
+#
+#  __init__.py
+#  pyblox -> util
+#
+#  By sbhadr (Sanjay Bhadra)
+#  Copyright © 2024- sbhadr (Sanjay Bhadra). All rights reserved.
+#
+
+# Some nifty methods that I like to keep handy because
+# I write them over and over for every project.
+# Bet you relate.. 
+#
+# Message from Sanjay 2024: Who the hell uses
+# the word nifty..?
+
+# Parent Class Modules
+from .req import Req
+from .io import io
+from .cached import Cached
