+++ src/webapi/util/cached.py
@@ -1,0 +1,12 @@
+#
+#  util -> cached.py
+#  pyblox
+#
+#  By sbhadr (Sanjay Bhadra)
+#  Copyright © 2024- sbhadr (Sanjay Bhadra). All rights reserved.
+#
+
+class Cached:
+	cache = {}
+	file_cache = {}
+	auth = {}
